requirements.txt
streamlit==1.55.0
openai==2.28.0
chromadb==1.5.5
pypdf==6.8.0
python-dateutil==2.9.0.post0
pytest==9.0.2
